+++ src/funcTest/groovy/info/solidsoft/gradle/pitest/functional/PitestPluginGeneralFunctionalSpec.groovy
@@ -113,6 +113,27 @@ class PitestPluginGeneralFunctionalSpec extends AbstractPitestFunctionalSpec {
             htmlFileWithReportForHelloPit.text.contains("System.out.println(&#34;Mutation to survive&#34;);")
     }
 
+    @Issue("https://github.com/szpak/gradle-pitest-plugin/issues/67")
+    void "reuses cached output"() {
+        given:
+            buildFile << getBasicGradlePitestConfig()
+        and:
+            writeHelloPitClass()
+            writeHelloPitTest()
+        when:
+            ExecutionResult result = runTasksSuccessfully('pitest', '--build-cache')
+            ExecutionResult result2 = runTasksSuccessfully('clean', 'pitest', '--build-cache')
+        then:
+            result.wasExecuted(':pitest')
+            result.getStandardOutput().contains("Build cache key for task ':pitest' is")
+//            //TODO: It's flaky - build cache for TestKit executions seems to be also cached
+//            //      Tests in Gradle itself have similar problem: https://github.com/gradle/gradle/blob/5ec3f672ed600a86280be490395d70b7bc634862/subprojects/core/src/integTest/groovy/org/gradle/api/tasks/CachedTaskIntegrationTest.groovy#L118-L132
+//            result.getStandardOutput().contains("Stored cache entry for task ':pitest'")
+        and:
+            result2.wasExecuted(':pitest')
+            result2.getStandardOutput().contains("Task :pitest FROM-CACHE")
+    }
+
     private String quoteBackslashesInWindowsPath(File file) {
         //There is problem with backslash within '' or "" while running this test on Windows: "unexpected char"
         return file.absolutePath.replaceAll('\\\\', '\\\\\\\\')