+++ src/main/groovy/info/solidsoft/gradle/pitest/PitestPlugin.groovy
@@ -52,9 +52,9 @@ class PitestPlugin implements Plugin<Project> {
     public final static String PITEST_REPORT_DIRECTORY_NAME = 'pitest'
     public final static String PITEST_CONFIGURATION_NAME = 'pitest'
 
-    public final static String DEFAULT_PITEST_VERSION = '1.6.3'
-    @Internal
-    public static final GradleVersion MINIMAL_SUPPORTED_GRADLE_VERSION = GradleVersion.version("5.6") //public as used also in regression tests
+    public final static String DEFAULT_PITEST_VERSION = '1.7.0'
+    @Internal   //6.4 due to main -> mainClass change to avoid deprecation warning in Gradle 7.x - https://github.com/szpak/gradle-pitest-plugin/pull/289
+    public static final GradleVersion MINIMAL_SUPPORTED_GRADLE_VERSION = GradleVersion.version("6.4") //public as used also in regression tests
 
     private static final String PITEST_JUNIT5_PLUGIN_NAME = "junit5"
     private final static List<String> DYNAMIC_LIBRARY_EXTENSIONS = ['so', 'dll', 'dylib']
@@ -157,7 +157,6 @@ class PitestPlugin implements Plugin<Project> {
         task.verbose.set(extension.verbose)
         task.timeoutFactor.set(extension.timeoutFactor)
         task.timeoutConstInMillis.set(extension.timeoutConstInMillis)
-        task.maxMutationsPerClass.set(extension.maxMutationsPerClass)
         task.childProcessJvmArgs.set(extension.jvmArgs)
         task.outputFormats.set(extension.outputFormats)
         task.failWhenNoMutations.set(extension.failWhenNoMutations)
@@ -175,8 +174,9 @@ class PitestPlugin implements Plugin<Project> {
         task.additionalClasspath.setFrom({
             List<FileCollection> testRuntimeClasspath = (extension.testSourceSets.get() as Set<SourceSet>)*.runtimeClasspath
             FileCollection combinedTaskClasspath = project.objects.fileCollection().from(testRuntimeClasspath)
+            List<String> fileExtensionsToFilter = extension.fileExtensionsToFilter.getOrElse([])
             FileCollection filteredCombinedTaskClasspath = combinedTaskClasspath.filter { File file ->
-                !extension.fileExtensionsToFilter.getOrElse([]).find { extension -> file.name.endsWith(".$extension") }
+                !fileExtensionsToFilter.find { extension -> file.name.endsWith(".$extension") }
             } + project.files(allMutableCodePaths)
             return filteredCombinedTaskClasspath
         } as Callable<FileCollection>)